+++ Deployer.Gui/ViewModels/Design/WimFileRequirementViewModelDesign.cs
@@ -1,8 +1,7 @@
 using System.Reactive;
 using ReactiveUI;
-using Spectre.Console.Cli;
 
-namespace Deployer.Gui
+namespace Deployer.Gui.ViewModels.Design
 {
     public class WimFileRequirementViewModelDesign
     {
